Create Calendar Quiz.py
Calendar Quiz
 
Super Quiz Bee is a famous quiz Competition that tests students on a wide variety of academic subjects. This week’s participants were kids of age 12 to 15 and the quiz questions were based on Gregorian calendar.
 
In the first round of the competition, the Host of the event told the participants that it was Monday on the date 01/01/2001. Later he questioned each one of the participant what would be the day on the 1st January, giving them a particular year. Write a program to help the Host validate the answers given by the participants.
 
Input Format:
The first line contains an integer that corresponds to a year.

Output Format:
Output the day on the 1st January of that given year.
Refer sample input and output for formatting specifications.

Sample Input 1:
1994

Sample Output 1:
Saturday

Sample Input 2:
2014

Sample Output 2:
Wednesday