Create Charity Dinner.py
Charity Dinner
 
WeCanNGO Trust is organizing a charity dinner at St. John’s College. Since older students are both wiser and richer, the members of the trust decide that the price of each ticket will be based on how many years the students have been in the school. A first year student will buy a PINK ticket, a second year student will buy a GREEN ticket, a third year student will buy a RED ticket, and a fourth year student will buy an ORANGE ticket.

Assume that all tickets are sold. Each colour of ticket is uniquely priced. Write a program to output all combinations of tickets that produce exactly the desired amount to be raised. The combinations may appear in a specific order. First ORANGE is considered, then RED, then GREEN and finally PINK. Also display the total number of combinations found and the smallest number of tickets to be printed to raise the desired amount so that printing cost is minimized.

Input Format:
First 4 lines of the input correspond to the cost of a PINK, GREEN, RED, ORANGE ticket (in the exact order).
Last line of the input corresponds to the exact amount of money to be raised by selling tickets.

Output Format:
Output all combinations of tickets that produce exactly the desired amount to be raised. The combinations may appear in any order. Output the total number of combinations found. Output the smallest number of tickets to print to raise the desired amount so that printing cost is minimized.
Refer sample input and output for formatting specifications.

Sample Input 1:
1
2
3
4
3

Sample Output 1:
# of PINK is 0 # of GREEN is 0 # of RED is 1 # of ORANGE is 0
# of PINK is 1 # of GREEN is 1 # of RED is 0 # of ORANGE is 0
# of PINK is 3 # of GREEN is 0 # of RED is 0 # of ORANGE is 0
Total combinations is 3
Minimum number of tickets to print is 1

Sample Input 2:
1
2
4
3
4

Sample Output 2:
# of PINK is 0 # of GREEN is 0 # of RED is 1 # of ORANGE is 0
# of PINK is 0 # of GREEN is 2 # of RED is 0 # of ORANGE is 0
# of PINK is 1 # of GREEN is 0 # of RED is 0 # of ORANGE is 1
# of PINK is 2 # of GREEN is 1 # of RED is 0 # of ORANGE is 0
# of PINK is 4 # of GREEN is 0 # of RED is 0 # of ORANGE is 0
Total combinations is 5
Minimum number of tickets to print is 1