Create Recreation Fair.py
Recreation Fair
 
The Warners Club arranged for a Recreation Fair for kids and adults for the winter season ahead. The Club offered outdoor activities for adults like hiking, cycling, power soccer and indoor activities like chess, carom and card games for kids below 10 years. Gary and Dory accompanied their parents for the fair and wanted to play the cricket collectable card game as they are crazy about the collectable cards.

In their spare time, they usually have a way of playing some game involving such cards. Both also have the habit of exchanging the repeated cards with their friends. That day at the Fair, Gary and Dory thought about a different game. With the cards in hand, each person tried to make an exchange with the other person following this simple rule: each person must count how many cards he owned. After this, they had to split these cards into stacks, all of it with the same size, as large as it was possible for both players. Then, each one chooses one of the friend's card stacks to receive.

For example, if Gary and Dory would change the cards with respectively 8 and 12 cards each, both must put his cards in stacks of four cards (Gary would have two stacks and Dory had three stacks), and both choose a stack from his friend to receive it.

Write a program to help Gary and Dory find the maximum size of the stack of cards which can be exchanged between the two players.

C Function Specifications:
Use the function name and the argument as:
int findValue(f1,f2):
The function must return the maximum size of the stack of cards which can be exchanged between the two players.

Python Function Specifications:
Use the function name and the argument as:
def findValue(f1,f2):
The function must return the maximum size of the stack of cards which can be exchanged between the two players.

Input Format:
The first line of the input contain an integer F1 (1 ≤ F1 ≤ 1000), which corresponds to the amount of collectable cards that Gary have to change.
The second line of the input contain an integer F2 (1 ≤ F2 ≤ 1000), which corresponds to the amount of collectable cards that Dory have to change.

Output Format:
Output the maximum size of the stack of cards which can be exchanged between the two players, in a single line.
Refer sample input and output for formatting specifications.

Sample Input 1:
8
12
Sample Output 1:
4

Sample Input 2:
9
27
Sample Output 2:
9